Move VR scripts in separate folder
